refactor    : auth 프로젝트 구조 리펙터링, 예외처리 수정 - 프론트엔드 처리를 위해 토큰 유효기간 만료시 401에러와함께 메시지   반환하게 수정 - 프로젝트 구조 Auth서버 리펙토링
