+++ auth/src/main/java/com/_hateam/auth/infrastructure/repository/RefreshTokenRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
     Optional<RefreshToken> findByUserId(Long userId);
-    void deleteByUserId(Long userId);
+
 }
