+++ auth/src/main/java/com/_hateam/auth/presentation/controller/AuthController.java
@@ -13,6 +13,7 @@ public class AuthController {
 
     private final AuthService authService;
 
+    //로그인
     @PostMapping("/signin")
     public ResponseEntity<ResponseDto<UserSignInResDto>> verifyUser(@RequestBody UserSignInReqDto userSignInReqDto) {
         ResponseDto<UserSignInResDto> response = authService.authenticate(userSignInReqDto);
@@ -20,8 +21,7 @@ public class AuthController {
     }
 
 
-
-    // AuthController 수정
+    //리프레시 토큰 재발급
     @PostMapping("/refresh")
     public ResponseEntity<ResponseDto<UserSignInResDto>> refreshToken(@RequestBody RefreshReqDto refreshReqDto) {
 
@@ -38,14 +38,14 @@ public class AuthController {
         return ResponseEntity.ok(response);
     }
 
-
+    //로그아웃
     @PostMapping("/logout")
     public ResponseEntity<ResponseDto<Void>> logout(@RequestHeader("Authorization") String authHeader) {
         ResponseDto<Void> response = authService.logoutWithToken(authHeader);
         return ResponseEntity.ok(response);
     }
 
-
+    //유저삭제시 리프레시토큰도 삭제
     @PostMapping("/users/{userId}/tokens")
     public ResponseEntity<ResponseDto<Void>> deleteUserTokens(@PathVariable Long userId) {
         ResponseDto<Void> response = authService.deleteAllUserTokens(userId);
