+++ gateway/src/main/java/com/_hateam/gateway/domain/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com._hateam.config;
+package com._hateam.gateway.domain.enums;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
