+++ gateway/src/main/java/com/_hateam/gateway/infrastructure/config/WebConfig.java
@@ -1,0 +1,14 @@
+package com._hateam.gateway.infrastructure.config;
+
+import org.springframework.boot.autoconfigure.web.WebProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class WebConfig {
+
+    @Bean
+    public WebProperties.Resources resources() {
+        return new WebProperties.Resources();
+    }
+}
