+++ gateway/src/main/java/com/_hateam/gateway/infrastructure/filter/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com._hateam.filter;
+package com._hateam.gateway.infrastructure.filter;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
